move test to it's own file
